session undefined issue on production is solved
