+++ package.json
@@ -61,7 +61,7 @@
     "mongoose": "^8.7.0",
     "negotiator": "^1.0.0",
     "next": "14.2.13",
-    "next-auth": "^4.24.8",
+    "next-auth": "^5.0.0-beta.25",
     "next-intl": "^3.25.1",
     "next-nprogress-bar": "^2.4.3",
     "nodemailer": "^6.9.16",
@@ -98,7 +98,9 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
     "handlebars-loader": "^1.7.3",
+    "html-loader": "^5.1.0",
     "husky": "^9.1.6",
+    "ignore-loader": "^0.1.2",
     "lint-staged": "^15.2.10",
     "postcss": "^8",
     "prettier": "^3.3.3",
