+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,1 @@
-import { authOptions } from '@/lib/auth-config';
-import NextAuth from 'next-auth';
-
-const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST };
+export { GET, POST } from '@/auth';
