+++ src/lib/auth.ts
@@ -1,0 +1,7 @@
+import { auth } from '@/auth';
+
+export const currentUser = async () => {
+  const session = await auth();
+
+  return session?.user;
+};
