+++ src/middleware.ts
@@ -1,35 +1,8 @@
 // middleware.ts
 import { NextResponse, NextRequest } from 'next/server';
-import { match as matchLocale } from '@formatjs/intl-localematcher';
-import Negotiator from 'negotiator';
-import { Locale, i18n } from '../i18n.config';
-
-function getLocale(request: NextRequest): string | undefined {
-  const preferredLanguage = request.cookies.get('preferred-language')
-    ?.value as Locale;
-
-  if (preferredLanguage && i18n.locales.includes(preferredLanguage)) {
-    return preferredLanguage;
-  }
-
-  const negotiatorHeaders: Record<string, string> = {};
-  request.headers.forEach((value, key) => (negotiatorHeaders[key] = value));
-
-  // @ts-expect-error locales are readonly
-  const locales: string[] = i18n.locales;
-  const languages = new Negotiator({ headers: negotiatorHeaders }).languages();
-  const locale = matchLocale(languages, locales, i18n.defaultLocale);
-  return locale;
-}
-
-const protectedRoutes = [
-  '/customer/dashboard',
-  '/customer/categories',
-  '/customer/rules',
-  '/customer/expenses',
-  '/customer/write-off',
-  '/customer/settings',
-];
+import { i18n } from '../i18n.config';
+import { getLocale } from './utils/helpers/getLocals';
+import { protectedRoutes } from './routes';
 
 export async function middleware(request: NextRequest) {
   const pathname = request.nextUrl.pathname;
