+++ src/routes/index.ts
@@ -1,0 +1,8 @@
+export const protectedRoutes = [
+  '/customer/dashboard',
+  '/customer/categories',
+  '/customer/rules',
+  '/customer/expenses',
+  '/customer/write-off',
+  '/customer/settings',
+];
